requirements.txt
duckdb==0.7.1
boto3==1.26.117
forex-python==1.8
pandas==2.0.0
snowflake-connector-python==3.0.3
google-cloud-bigquery==3.10.0
google-cloud-secret-manager==2.10.0
pyarrow==11.0.0